requirements.txt
flask==3.1.0
flask-cors==5.0.1
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.3
requests==2.32.3
python-dotenv==1.0.1
gunicorn==23.0.0
joblib==1.4.2
